requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
click==8.1.2
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.1
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
Mako==1.2.0
MarkupSafe==2.1.1
poetry==1.1.12
psycopg2==2.9.3
pydantic==1.9.0
python-dotenv==0.20.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.6
